# Removed defered mutex unlocks in favor of explicit calls