@mjs 2016-09-14
Nice. I guess we need to balance performance against the increased risk of problems when someone updates the code and forgets to Unlock.

For `available` which is really short I have no problem with avoiding defer. But for the other methods I'm a bit more wary. I guess if there's tests which ensure that the mutex is always unlocked then it might be ok,

How much is the performance hit of the defers hurting you in actual use?