@curlymon 2016-11-27
Sorry for the extreme delay in response.

My current need for the speed increase entailed here has changed since I originally wrote this and I no longer need this specific change. That and with the speed up of defer coming in 1.8 I think that the trade off of performance and safety are much more in favor of safety so I will close this set of changes out.